requirements.txt
google-generativeai
langchain
langchain-core
langchain-google-genai
langchainplus-sdk
langchain-ollama
langchain-openai
langchain-text-splitters
python-dotenv
